🤖 Move re-run button into Work Details card with header and link styling
