+++ platform/scms/app/routes/app/works.$workId.upload.$workVersionId/metadata-extract/MetadataFormCard.tsx
@@ -1,3 +1,4 @@
+import { Bot, RefreshCw } from 'lucide-react';
 import { ui, LoadingSpinner } from '@curvenote/scms-core';
 import type { ExtractedMetadata } from './anthropic.server';
 import { WorkTitleForm } from '../WorkTitleForm';
@@ -11,12 +12,21 @@ function authorsFromExtracted(extracted: ExtractedMetadata | null): string {
     .join(', ');
 }
 
+/** Shorten a file name for the re-run label. */
+function shortenFileName(name: string, max = 32): string {
+  if (name.length <= max) return name;
+  return `${name.slice(0, max - 1).trimEnd()}…`;
+}
+
 export interface MetadataFormCardProps {
   extractedMetadata: ExtractedMetadata | null;
   /** True while an extraction request is in flight (drives the overlay). */
   isExtractingMetadata: boolean;
   title: string;
   authors: string;
+  /** Name of the file the re-run control targets; hides the control when undefined. */
+  reRunFileName?: string;
+  onReRunExtraction?: () => void;
 }
 
 export function MetadataFormCard({
@@ -24,6 +34,8 @@ export function MetadataFormCard({
   isExtractingMetadata,
   title,
   authors,
+  reRunFileName,
+  onReRunExtraction,
 }: MetadataFormCardProps) {
   const displayTitle = (title?.trim() ? title : extractedMetadata?.title) ?? '';
   const initialAuthors = authors?.trim() ? authors : authorsFromExtracted(extractedMetadata);
@@ -40,6 +52,10 @@ export function MetadataFormCard({
           <p className="text-sm text-muted-foreground">Extracting metadata…</p>
         </div>
       )}
+      <div className="flex gap-2 items-center">
+        <Bot className="w-5 h-5 text-muted-foreground" />
+        <h3 className="text-base font-semibold">Work Details</h3>
+      </div>
       <WorkTitleForm title={displayTitle} />
       <AuthorsForm initialAuthors={initialAuthors} />
       {extractedMetadata != null && (
@@ -50,6 +66,22 @@ export function MetadataFormCard({
           </pre>
         </details>
       )}
+      {reRunFileName && onReRunExtraction ? (
+        <div className="flex justify-end">
+          <ui.Button
+            type="button"
+            variant="link"
+            size="sm"
+            className="p-0 h-auto"
+            onClick={onReRunExtraction}
+            disabled={isExtractingMetadata}
+            title={`Re-run extraction on ${reRunFileName}`}
+          >
+            <RefreshCw className="mr-1.5 w-3.5 h-3.5" />
+            {`re-run extraction on ${shortenFileName(reRunFileName)}`}
+          </ui.Button>
+        </div>
+      ) : null}
     </ui.Card>
   );
 }
